Organised Imports and Format All Codes

+++ src/main/java/iss/sa46team12/springclub/controllers/UserController.java
@@ -9,6 +9,7 @@ import java.util.concurrent.TimeUnit;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 import iss.sa46team12.springclub.email.SendEmail;
 import iss.sa46team12.springclub.models.BookingDetails;
 import iss.sa46team12.springclub.models.Bookings;
@@ -66,9 +68,8 @@ public class UserController {
 	public ModelAndView editProfile(HttpSession session) {
 		int userid;
 		try {
-			userid = (int) session.getAttribute("UserID"); 
-		}
-		catch(NullPointerException e) {
+			userid = (int) session.getAttribute("UserID");
+		} catch (NullPointerException e) {
 			userid = 1;
 		}
 		ModelAndView mav = new ModelAndView("profile", "password", new Password());
@@ -102,7 +103,7 @@ public class UserController {
 		c2.setTime(new Date());// today date
 		long dayleft = daysBetween(c, c2);
 		mav.addObject("dayleft", dayleft);
-		double dayleftint = (double) dayleft;
+		double dayleftint = dayleft;
 		// Membership Date Percentage
 		double days = 0; // package year in days
 		double percentage = 0;
@@ -155,9 +156,8 @@ public class UserController {
 
 		int userid;
 		try {
-			userid = (int) session.getAttribute("UserID"); 
-		}
-		catch(NullPointerException e) {
+			userid = (int) session.getAttribute("UserID");
+		} catch (NullPointerException e) {
 			userid = 1;
 		}
 		if (result.hasErrors()) {
@@ -194,9 +194,8 @@ public class UserController {
 
 		int userid;
 		try {
-			userid = (int) session.getAttribute("UserID"); 
-		}
-		catch(NullPointerException e) {
+			userid = (int) session.getAttribute("UserID");
+		} catch (NullPointerException e) {
 			userid = 1;
 		}
 		if (result.hasErrors()) {
@@ -223,9 +222,8 @@ public class UserController {
 
 		int userid;
 		try {
-			userid = (int) session.getAttribute("UserID"); 
-		}
-		catch(NullPointerException e) {
+			userid = (int) session.getAttribute("UserID");
+		} catch (NullPointerException e) {
 			userid = 1;
 		}
 		ModelAndView mav;
@@ -277,9 +275,8 @@ public class UserController {
 
 		int userid;
 		try {
-			userid = (int) session.getAttribute("UserID"); 
-		}
-		catch(NullPointerException e) {
+			userid = (int) session.getAttribute("UserID");
+		} catch (NullPointerException e) {
 			userid = 1;
 		}
 		String bookingStatus = "CONFIRMED";
